Renaming and refactoring to better reflect terms used in the report.
