+++ checker/src/main/java/module-info.java
@@ -8,7 +8,6 @@ module chortlin.checker {
     exports com.dickow.chortlin.checker.correlation.builder;
     exports com.dickow.chortlin.checker.choreography;
     exports com.dickow.chortlin.checker.choreography.participant;
-    exports com.dickow.chortlin.checker.choreography.method;
     exports com.dickow.chortlin.checker.checker;
     exports com.dickow.chortlin.checker.checker.factory;
     exports com.dickow.chortlin.checker.checker.session;
