+++ checker/src/main/kotlin/com/dickow/chortlin/checker/ast/ASTVisitor.kt
@@ -6,6 +6,5 @@ interface ASTVisitor {
     fun visitEnd(astNode: End)
     fun visitInteraction(astNode: Interaction)
     fun visitReturnFrom(astNode: ReturnFrom)
-    fun visitParallel(astNode: Parallel)
     fun visitChoice(astNode: Choice)
 }
