+++ checker/src/main/kotlin/com/dickow/chortlin/checker/ast/types/ASTNode.kt
@@ -6,7 +6,6 @@ import com.dickow.chortlin.checker.ast.Label
 import com.dickow.chortlin.checker.ast.types.placeholder.Placeholder
 import com.dickow.chortlin.checker.checker.SatisfactionRelationship
 import com.dickow.chortlin.checker.choreography.Choreography
-import com.dickow.chortlin.checker.choreography.method.ObservableMethod
 import com.dickow.chortlin.checker.choreography.participant.ObservableParticipant
 import com.dickow.chortlin.checker.choreography.participant.Participant
 
@@ -22,24 +21,14 @@ abstract class ASTNode(var previous: ASTNode?, var next: ASTNode?) : ASTBuilder,
         return build()
     }
 
-    override fun parallel(path: (ASTBuilder) -> Choreography): ASTBuilder {
-        val parallelPath = path(Choreography.builder())
-        val next = Parallel(parallelPath, this, null)
-        parallelPath.start.previous = next
+    override fun returnFrom(receiver: ObservableParticipant, label: String): ASTBuilder {
+        val next = ReturnFrom(receiver, Label(label), this, null)
         this.next = next
         return next
     }
 
-    override fun returnFrom(observableMethod: ObservableMethod, label: String): ASTBuilder {
-        val observableReceiver = ObservableParticipant(observableMethod.participant.identifier, observableMethod.method)
-        val next = ReturnFrom(observableReceiver, Label(label), this, null)
-        this.next = next
-        return next
-    }
-
-    override fun interaction(sender: Participant, observableMethod: ObservableMethod, label: String): ASTBuilder {
-        val observableReceiver = ObservableParticipant(observableMethod.participant.identifier, observableMethod.method)
-        val next = Interaction(sender, observableReceiver, Label(label), this, null)
+    override fun interaction(sender: Participant, receiver: ObservableParticipant, label: String): ASTBuilder {
+        val next = Interaction(sender, receiver, Label(label), this, null)
         this.next = next
         return next
     }
