--- checker/src/main/kotlin/com/dickow/chortlin/checker/choreography/method/MethodFactory.kt
@@ -1,9 +1,0 @@
-package com.dickow.chortlin.checker.choreography.method
-
-import com.dickow.chortlin.checker.choreography.participant.InternalParticipant
-
-object MethodFactory {
-    fun method(participant: InternalParticipant, methodName: String): ObservableMethod {
-        return ObservableMethod(methodName, participant)
-    }
-}
