--- checker/src/main/kotlin/com/dickow/chortlin/checker/choreography/method/ObservableMethod.kt
@@ -1,5 +1,0 @@
-package com.dickow.chortlin.checker.choreography.method
-
-import com.dickow.chortlin.checker.choreography.participant.InternalParticipant
-
-data class ObservableMethod(val method: String, val participant: InternalParticipant)
