+++ checker/src/main/kotlin/com/dickow/chortlin/checker/choreography/participant/ObservableParticipant.kt
@@ -2,4 +2,4 @@ package com.dickow.chortlin.checker.choreography.participant
 
 import com.dickow.chortlin.shared.observation.Observable
 
-class ObservableParticipant(canonicalClassName: String, method: String) : Observable(canonicalClassName, method)
+class ObservableParticipant(identifier: String, method: String) : Observable(identifier, method), Participant
