+++ checker/src/main/kotlin/com/dickow/chortlin/checker/correlation/builder/CorrelationBuilder.kt
@@ -1,6 +1,5 @@
 package com.dickow.chortlin.checker.correlation.builder
 
-import com.dickow.chortlin.checker.choreography.method.ObservableMethod
 import com.dickow.chortlin.checker.choreography.participant.ObservableParticipant
 import com.dickow.chortlin.checker.correlation.Correlation
 import com.dickow.chortlin.checker.correlation.CorrelationFunction
@@ -9,20 +8,18 @@ import com.dickow.chortlin.checker.correlation.ReturnTypesFunction
 import com.dickow.chortlin.checker.correlation.path.Path
 import java.util.*
 
-class CorrelationBuilder(private val observableMethod: ObservableMethod,
+class CorrelationBuilder(private val observable: ObservableParticipant,
                          private val correlationFunction: CorrelationFunction)
 {
     private val inputFunctions: MutableList<InputTypesFunction> = LinkedList()
     private val returnFunctions: MutableList<ReturnTypesFunction> = LinkedList()
 
     fun done(): Correlation {
-        val observableParticipant = ObservableParticipant(observableMethod.participant.identifier, observableMethod.method)
-        return Correlation(observableParticipant, correlationFunction, inputFunctions, returnFunctions)
+        return Correlation(observable, correlationFunction, inputFunctions, returnFunctions)
     }
 
     fun noExtensions(): Correlation {
-        val observableParticipant = ObservableParticipant(observableMethod.participant.identifier, observableMethod.method)
-        return Correlation(observableParticipant, correlationFunction, emptyList(), emptyList())
+        return Correlation(observable, correlationFunction, emptyList(), emptyList())
     }
 
     fun extendFromInput(key: String, path: Path): CorrelationBuilder {
