+++ checker/src/main/kotlin/com/dickow/chortlin/checker/correlation/factory/CorrelationFactory.kt
@@ -2,7 +2,7 @@
 
 package com.dickow.chortlin.checker.correlation.factory
 
-import com.dickow.chortlin.checker.choreography.method.ObservableMethod
+import com.dickow.chortlin.checker.choreography.participant.ObservableParticipant
 import com.dickow.chortlin.checker.correlation.CorrelationFunction
 import com.dickow.chortlin.checker.correlation.builder.CorrelationBuilder
 import com.dickow.chortlin.checker.correlation.builder.CorrelationParticipantMappingBuilder
@@ -16,7 +16,7 @@ object CorrelationFactory {
     }
 
     @JvmStatic
-    fun correlation(method: ObservableMethod, key: String, correlationFunction: Path): CorrelationBuilder {
-        return CorrelationBuilder(method, CorrelationFunction(key, correlationFunction))
+    fun correlation(observable: ObservableParticipant, key: String, correlationFunction: Path): CorrelationBuilder {
+        return CorrelationBuilder(observable, CorrelationFunction(key, correlationFunction))
     }
 }
