+++ interception/src/main/java/module-info.java
@@ -11,11 +11,11 @@ module chortlin.interception {
 
     exports com.dickow.chortlin.interception;
     exports com.dickow.chortlin.interception.configuration;
-    exports com.dickow.chortlin.interception.defaults;
+    exports com.dickow.chortlin.interception.configuration.defaults;
     exports com.dickow.chortlin.interception.instrumentation;
     exports com.dickow.chortlin.interception.instrumentation.advice;
     exports com.dickow.chortlin.interception.sending;
     exports com.dickow.chortlin.interception.annotations;
-    exports com.dickow.chortlin.interception.dto;
+    exports com.dickow.chortlin.interception.sending.dto;
     exports com.dickow.chortlin.interception.observation;
 }
