+++ interception/src/main/kotlin/com/dickow/chortlin/interception/configuration/defaults/HttpSender.kt
@@ -1,4 +1,4 @@
-package com.dickow.chortlin.interception.defaults
+package com.dickow.chortlin.interception.configuration.defaults
 
 import com.dickow.chortlin.interception.configuration.HttpConfiguration
 import com.dickow.chortlin.interception.sending.TraceSender
